Update unity switch mode combobox
