+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/ui/UnitTest/UseUnityPlayLauncherAction.kt
@@ -1,0 +1,25 @@
+package com.jetbrains.rider.plugins.unity.ui.UnitTest
+
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.project.DumbAwareAction
+import com.jetbrains.rider.model.UnitTestLaunchPreference
+import com.jetbrains.rider.model.rdUnityModel
+import com.jetbrains.rider.plugins.unity.isConnectedToEditor
+import com.jetbrains.rider.projectView.solution
+
+class UseUnityPlayLauncherAction : DumbAwareAction(PlayModeDescription, "Run with Unity Editor in Play Mode", null) {
+    companion object {
+        const val PlayModeDescription = "Unity Editor - Play Mode"
+    }
+
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+        project.solution.rdUnityModel.unitTestPreference.value = UnitTestLaunchPreference.PlayMode
+    }
+
+    override fun update(e: AnActionEvent) {
+        val project = e.project ?: return
+        e.presentation.isEnabled = project.isConnectedToEditor()
+        e.presentation.isVisible = true
+    }
+}
